Split large component into two smaller ones.
